src/requirements.txt
fastapi==0.115.6
python-dotenv==1.0.0
uvicorn==0.23.1
ruff==0.7.3
pymongo==4.10.1
faker==33.1.0
httpx==0.28.0